requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnx==1.17.0
onnxruntime==1.20.1
openai==0.28.0
ordered-set==4.1.0
packaging==24.2
propcache==0.2.0
protobuf==5.29.0
py2wasm==2.6.2
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyreadline3==3.5.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
zstandard==0.23.0
